new data and graph added to home

+++ app/src/main/java/com/example/appoftheyear2/Addgame.java
@@ -75,7 +75,7 @@ public class Addgame extends AppCompatActivity implements AdapterView.OnItemSele
             @Override
             public void onClick(View v) {
                 if (!nameInput.getText().toString().equals("")) {
-                    playtimeTotalHours = Integer.valueOf(hoursInput.getText().toString()) + (Integer.valueOf(minutesInput.getText().toString())/60) + ((Integer.valueOf(secondsInput.getText().toString())/60)/60);
+                    playtimeTotalHours = Float.valueOf(hoursInput.getText().toString()) + Float.valueOf((Float.valueOf(minutesInput.getText().toString())/60)) + Float.valueOf(((Float.valueOf(secondsInput.getText().toString())/60)/60));
                     Intent returnIntent = new Intent();
                     returnIntent.putExtra("nameInput", nameInput.getText().toString());
                     returnIntent.putExtra("genreInput", genreSelection);
